+++ lib/src/switch/switch.dart
@@ -1,2 +1,4 @@
 export 'sbb_switch.dart';
 export 'sbb_switch_list_item.dart';
+export 'theme/sbb_switch_style.dart';
+export 'theme/sbb_switch_theme_data.dart';