+++ CHANGELOG.md
@@ -24,6 +24,8 @@ It is expected that you keep this format strictly, since we depend on it in our 
 - added `style` to `SBBRadio`
 - added `SBBRadioThemeData` to override `SBBRadioStyle` within the `SBBTheme`
 - added `focusNode` & `autofocus` to `SBBRadio`
+- added `style` to `SBBSwitch` - see [v5 migration guide]
+- added `focusNode` & `autofocus` to `SBBSwitch`
 
 ### Changed
 
@@ -40,6 +42,7 @@ It is expected that you keep this format strictly, since we depend on it in our 
 - BREAKING: `SBBRadio` needs an `SBBRadioGroup` ancestor for its group value and `onChanged` callback - see [v5 migration guide]
 - dropped support for Flutter SDK 3.32.x
 - BREAKING: removed default left margin on component `SBBTextField` (#402)
+- BREAKING: `SBBSwitch` is now styled with own `SBBSwitchStyle` from `SBBSwitchThemeData` - see [v5 migration guide]
 
 ### Removed
 
@@ -105,7 +108,7 @@ It is expected that you keep this format strictly, since we depend on it in our 
 ### Added
 
 - added `SBBSliverFloatingHeaderbox` that allows the headerbox to expand and contract as the user scrolls.
-- added functional colors (e.g. `SBBColors.error`) and dark variants of additionalColors (e.g. `SBBColors.skyDark`)
+- added functional colors (e.g. `SBBColors.error`) and dark variants of `additionalColors` (e.g. `SBBColors.skyDark`)
 - added `errorColor` to `SBBBaseStyle`
 - added `systemOverlayStyle` to `SBBHeader`